# Add minimal Optimizer Transformer Config