--- a/packages/optimizer/index.js
+++ b/packages/optimizer/index.js
@@ -19,6 +19,7 @@ const {
   DomTransformer,
   TRANSFORMATIONS_AMP_FIRST,
   TRANSFORMATIONS_PAIRED_AMP,
+  TRANSFORMATIONS_MINIMAL,
   DEFAULT_CONFIG,
 } = require('./lib/DomTransformer.js');
 const NodeUtils = require('./lib/NodeUtils.js');
@@ -26,6 +27,7 @@ const NodeUtils = require('./lib/NodeUtils.js');
 module.exports = {
   create: (config = DEFAULT_CONFIG) => new DomTransformer(config),
   TRANSFORMATIONS_AMP_FIRST,
+  TRANSFORMATIONS_MINIMAL,
   TRANSFORMATIONS_PAIRED_AMP,
   DEFAULT_CONFIG,
   NodeUtils,