--- a/packages/optimizer/spec/end-to-end/body-only/expected_output.minimal.html
+++ b/packages/optimizer/spec/end-to-end/body-only/expected_output.minimal.html
@@ -1,0 +1,19 @@
+<!doctype html>
+<html amp i-amphtml-layout i-amphtml-no-boilerplate transformed="self;v=1">
+<head>
+  <meta data-auto charset="utf-8"><style amp-runtime i-amphtml-version="123456789000000">/* ampproject.org/rtv v0.css */amp-img[i-amphtml-ssr]:not(.i-amphtml-element):not([layout=container])>*{display: block;}</style>
+  <meta data-auto name="viewport" content="width=device-width,minimum-scale=1,initial-scale=1">
+  <link as="script" crossorigin="anonymous" href="https://cdn.ampproject.org/v0.mjs" rel="modulepreload">
+  <script async nomodule src="https://cdn.ampproject.org/v0.js"></script>
+  <script data-auto async src="https://cdn.ampproject.org/v0.mjs" type="module" crossorigin="anonymous"></script>
+  <script async nomodule src="https://cdn.ampproject.org/v0/amp-video-0.1.js" custom-element="amp-video"></script>
+  <script async src="https://cdn.ampproject.org/v0/amp-video-0.1.mjs" custom-element="amp-video" type="module" crossorigin="anonymous"></script>
+  <link data-auto rel="canonical" href=".">
+</head>
+<body>
+  <h1>Hello, AMP world!</h1>
+  <amp-video width="480" height="270" src="https://amp.dev/static/samples/video/tokyo.mp4" poster="https://amp.dev/static/samples/img/tokyo.jpg" layout="responsive" controls class="i-amphtml-layout-responsive i-amphtml-layout-size-defined" i-amphtml-layout="responsive">
+    <i-amphtml-sizer style="display:block;padding-top:56.25%"></i-amphtml-sizer>
+  </amp-video>
+</body>
+</html>