--- a/packages/optimizer/spec/end-to-end/trailing-template/expected_output.minimal.html
+++ b/packages/optimizer/spec/end-to-end/trailing-template/expected_output.minimal.html
@@ -1,0 +1,16 @@
+<!doctype html>
+<html ⚡ i-amphtml-layout i-amphtml-no-boilerplate transformed="self;v=1">
+<head>
+  <meta charset="utf-8"><style amp-runtime i-amphtml-version="123456789000000">/* ampproject.org/rtv v0.css */amp-img[i-amphtml-ssr]:not(.i-amphtml-element):not([layout=container])>*{display: block;}</style>
+  <meta name="viewport" content="width=device-width,minimum-scale=1">
+  <link as="script" crossorigin="anonymous" href="https://cdn.ampproject.org/v0.mjs" rel="modulepreload">
+  <script async nomodule src="https://cdn.ampproject.org/v0.js"></script>
+  <script async src="https://cdn.ampproject.org/v0.mjs" type="module" crossorigin="anonymous"></script>
+  <script async nomodule src="https://cdn.ampproject.org/v0/amp-mustache-0.2.js" custom-template="amp-mustache"></script>
+  <script async custom-template="amp-mustache" src="https://cdn.ampproject.org/v0/amp-mustache-0.2.mjs" type="module" crossorigin="anonymous"></script>
+  <link rel="canonical" href="self.html">
+</head>
+<body>
+  <template type="amp-mustache"></template>
+</body>
+</html>