… packages/optimizer/demo/benchmark/samples/amp.dev.html