* Implement a minimal transformer config executing only essential
  transformations.
* Add a benchmark comparing both versions

Results:

```
minimal x 147 ops/sec ±38.65% (72 runs sampled) 6.816074326178452ms
default x 6.93 ops/sec ±11.41% (40 runs sampled) 144.290958325ms
Fastest is minimal
```