@sebastianbenz 2021-03-30
Traced the individual transformer execution times:

```
AddMandatoryTags: 0.465ms
Markdown: 0.055ms
AutoExtensionImporter: 68.673ms
OptimizeImages: 0.059ms
OptimizeHeroImage: 1.654ms
OptimizeAmpBind: 0.071ms
ServerSideRendering: 3.025ms
AmpBoilerplateTransformer: 0.116ms
ReorderHeadTransformer: 0.429ms
RewriteAmpUrls: 0.497ms
GoogleFontsPreconnect: 0.107ms
PruneDuplicateResourceHints: 0.165ms
AddBlurryImagePlaceholders: 0.085ms
SeparateKeyframes: 317.038ms
AddTransformedFlag: 0.042ms
MinifyHtml: 6.029ms
AmpScriptCsp: 0.631ms
```

Maybe we should get rid of SeparateKeyframes...since we've raised the CSS limit it's less of a concern.